refactor: move animated-text to shared ui
