+++ apps/billets-web/shared/ui/animated-text/animated-text.tsx
