+++ apps/billets-web/shared/ui/animated-text/index.ts
@@ -1,0 +1,1 @@
+export { AnimatedText } from './animated-text';
